Use a separate integer type for the primary indices.
This allows the function to automatically choose a smaller integer type
for the secondary indices (i.e., the ones that make their way into the
final matrix), even if the primary dimension extent is large.
Conversely, we can use a smaller type to hold the primary indices when
only the secondary dimension extent is large.